Fix VapoursynthDecoder does not apply variables (#4)
Resolves #3

When `from_file` and `from_script` are evaluated, they are evaluated
before any variables can be set and there is no mechanism to reevaluate
(not necessary anyway) after setting variables.

Add optional variables parameter to from_file and from_script which will
be set in the initialization *before* the script or file is evaluated.

This will require a minor version bump to be propagated downstream.

Thanks,
\- Boats M.

---------

Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>